Move all trackalert code to parent and use symlinks in trackalert dir
